Stats_Service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
clickhouse-connect==0.7.0
confluent-kafka==2.3.0
protobuf==4.25.1
